Refactor sessions screen state
