+++ src/web/features/control-plane/hooks/useSessionsScreenState.ts
@@ -1,54 +1,18 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import type { TraceEvent } from '../../../../core/types';
-import {
-  continueChatSession,
-  createChatSession,
-  fetchChatSessionDetail,
-  fetchChatTurnReview,
-  fetchPendingSessionApproval,
-  fetchSessionRunningState,
-  resolvePendingSessionApproval,
-  sendChatSessionPrompt,
-  subscribeToChatSessionEvents,
-  cancelChatSession,
-  updateChatSessionSettings,
-  type ChatSessionDetail,
-  type ChatTurnReview,
-  type ControlPlaneState,
-  type PendingSessionApproval,
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import type {
+  ChatSessionDetail,
+  ChatTurnReview,
+  ControlPlaneState,
+  PendingSessionApproval,
 } from '../../../lib/api';
 import type { ToastInput } from '../../../components/ui/use-toast';
+import type { SessionsScreenState } from './sessions-screen/sessionStateTypes';
+import { useLiveSessionMessages } from './sessions-screen/useLiveSessionMessages';
+import { useSessionDetailSubscription } from './sessions-screen/useSessionDetailSubscription';
+import { useSessionMutations } from './sessions-screen/useSessionMutations';
+import { useSessionTurnReview } from './sessions-screen/useSessionTurnReview';
 
-export type SessionDetailValue = Exclude<ChatSessionDetail, null>;
-export type SessionTurn = SessionDetailValue['turns'][number];
-
-export type SessionsScreenState = {
-  activeSession?: ControlPlaneState['sessions'][number];
-  selectedSessionId?: string;
-  setSelectedSessionId: (sessionId: string) => void;
-  sessionDetail: ChatSessionDetail | null;
-  sessionDetailLoading: boolean;
-  sessionDetailError?: string;
-  sendingPrompt: boolean;
-  runInFlight: boolean;
-  memoryUpdating: boolean;
-  sendPromptError?: string;
-  sendPrompt: (prompt: string) => Promise<void>;
-  creatingSession: boolean;
-  sessionNotice?: string;
-  createSession: () => Promise<void>;
-  continueSession: () => Promise<void>;
-  cancelSessionRun: () => Promise<void>;
-  updateSessionSettings: (settings: { model?: string; driftEnabled?: boolean }) => Promise<void>;
-  pendingApproval: PendingSessionApproval;
-  resolveApproval: (approved: boolean) => Promise<void>;
-  selectedTurnId?: string;
-  setSelectedTurnId: (turnId: string) => void;
-  selectedTurn?: SessionTurn;
-  turnReview: ChatTurnReview | null;
-  turnReviewLoading: boolean;
-  turnReviewError?: string;
-};
+export type { SessionDetailValue, SessionsScreenState, SessionTurn } from './sessions-screen/sessionStateTypes';
 
 export function useSessionsScreenState(
   sessions: ControlPlaneState['sessions'] | undefined,
@@ -76,119 +40,9 @@ export function useSessionsScreenState(
   const [creatingSession, setCreatingSession] = useState(false);
   const [sessionNotice, setSessionNotice] = useState<string | undefined>();
   const selectedSessionId = options?.selectedSessionId ?? internalSelectedSessionId;
-  const setSelectedSessionId = options?.onSelectedSessionIdChange ?? setInternalSelectedSessionId;
+  const setRawSelectedSessionId = options?.onSelectedSessionIdChange ?? setInternalSelectedSessionId;
   const autoSelectSession = options?.autoSelectSession ?? true;
-  const onSessionsChangedRef = useRef(onSessionsChanged);
-
-  useEffect(() => {
-    onSessionsChangedRef.current = onSessionsChanged;
-  }, [onSessionsChanged]);
-
-  const appendPendingUserTurn = useCallback((prompt: string) => {
-    setSessionDetail((current) => {
-      if (!current) {
-        return current;
-      }
-
-      const nextMessages = current.messages.filter((message) => (
-        message.id !== 'live-user'
-        && message.id !== 'live-assistant'
-        && message.id !== 'live-run-status'
-      ));
-      return {
-        ...current,
-        messages: [
-          ...nextMessages,
-          {
-            id: 'live-user',
-            role: 'user',
-            text: prompt,
-          },
-          {
-            id: 'live-run-status',
-            role: 'assistant',
-            text: 'Heddle is working…',
-            isStreaming: true,
-            isPending: true,
-          },
-        ],
-      };
-    });
-  }, []);
-
-  const upsertLiveStatusMessage = useCallback((id: string, text: string, options?: { pending?: boolean; streaming?: boolean }) => {
-    setSessionDetail((current) => {
-      if (!current) {
-        return current;
-      }
-
-      const nextMessages = [...current.messages];
-      const existingIndex = nextMessages.findIndex((message) => message.id === id);
-      const nextMessage = {
-        id,
-        role: 'assistant' as const,
-        text,
-        isPending: options?.pending,
-        isStreaming: options?.streaming,
-      };
-
-      if (existingIndex >= 0) {
-        nextMessages[existingIndex] = nextMessage;
-      } else {
-        nextMessages.push(nextMessage);
-      }
-
-      return {
-        ...current,
-        messages: nextMessages,
-      };
-    });
-  }, []);
-
-  const removeLiveStatusMessage = useCallback((id: string) => {
-    setSessionDetail((current) => {
-      if (!current) {
-        return current;
-      }
-
-      return {
-        ...current,
-        messages: current.messages.filter((message) => message.id !== id),
-      };
-    });
-  }, []);
-
-  const upsertLiveAssistantMessage = useCallback((text: string, isDone: boolean) => {
-    setSessionDetail((current) => {
-      if (!current) {
-        return current;
-      }
-
-      const nextMessages = current.messages.filter((message) => message.id !== 'live-run-status');
-      const lastMessage = nextMessages.at(-1);
-      if (lastMessage?.id === 'live-assistant' && lastMessage.role === 'assistant') {
-        nextMessages[nextMessages.length - 1] = {
-          ...lastMessage,
-          text,
-          isStreaming: !isDone,
-          isPending: !isDone,
-        };
-      } else {
-        nextMessages.push({
-          id: 'live-assistant',
-          role: 'assistant',
-          text,
-          isStreaming: !isDone,
-          isPending: !isDone,
-        });
-      }
-
-      return {
-        ...current,
-        messages: nextMessages,
-      };
-    });
-  }, []);
+  const liveMessages = useLiveSessionMessages(setSessionDetail);
 
   useEffect(() => {
     if (!autoSelectSession) {
@@ -199,7 +53,7 @@ export function useSessionsScreenState(
       if (selectedSessionId && sessionDetail?.id === selectedSessionId) {
         return;
       }
-      setSelectedSessionId(undefined);
+      setRawSelectedSessionId(undefined);
       return;
     }
 
@@ -208,483 +62,77 @@ export function useSessionsScreenState(
     }
 
     if (!selectedSessionId || !sessions.some((session) => session.id === selectedSessionId)) {
-      setSelectedSessionId(sessions[0].id);
+      setRawSelectedSessionId(sessions[0].id);
     }
-  }, [autoSelectSession, selectedSessionId, sessionDetail?.id, sessions, setSelectedSessionId]);
+  }, [autoSelectSession, selectedSessionId, sessionDetail?.id, sessions, setRawSelectedSessionId]);
 
   const selectSession = useCallback((sessionId: string) => {
-    setSelectedSessionId(sessionId);
+    setRawSelectedSessionId(sessionId);
     setSelectedTurnId(undefined);
     setSessionNotice(undefined);
     setSendPromptError(undefined);
-  }, [setSelectedSessionId]);
+  }, [setRawSelectedSessionId]);
 
-  useEffect(() => {
-    if (!selectedSessionId) {
-      setSessionDetail(null);
-      setSessionDetailError(undefined);
-      return;
-    }
-
-    const sessionId = selectedSessionId;
-    let cancelled = false;
-    let sessionUpdateTimeout: number | undefined;
-
-    async function refresh(options: { silent?: boolean } = {}) {
-      if (!options.silent) {
-        setSessionDetailLoading(true);
-      }
-      try {
-        const next = await fetchChatSessionDetail(sessionId);
-        if (!cancelled) {
-          setSessionDetail((current) => options.silent ? mergeTransientMessages(current, next) : next);
-          setSessionDetailError(undefined);
-        }
-      } catch (refreshError) {
-        if (!cancelled) {
-          setSessionDetailError(refreshError instanceof Error ? refreshError.message : String(refreshError));
-        }
-      } finally {
-        if (!cancelled && !options.silent) {
-          setSessionDetailLoading(false);
-        }
-      }
-    }
-
-    void refresh();
-    void fetchSessionRunningState(sessionId).then((state) => setRunInFlight(state.running));
-    const unsubscribe = subscribeToChatSessionEvents(sessionId, (event) => {
-      if (event.type === 'session.updated') {
-        if (sessionUpdateTimeout !== undefined) {
-          window.clearTimeout(sessionUpdateTimeout);
-        }
-        sessionUpdateTimeout = window.setTimeout(() => {
-          void fetchSessionRunningState(sessionId).then((state) => {
-            if (!cancelled) {
-              setRunInFlight(state.running);
-            }
-          });
-          void refresh({ silent: true });
-          onSessionsChangedRef.current?.();
-        }, 300);
-        return;
-      }
-
-      if (event.type !== 'session.event' || !event.event || typeof event.event !== 'object') {
-        return;
-      }
-
-      const liveEvent = event.event as {
-        type?: string;
-        text?: string;
-        done?: boolean;
-        tool?: string;
-        step?: number;
-        durationMs?: number;
-        outcome?: string;
-        summary?: string;
-        event?: TraceEvent;
-        status?: 'running' | 'finished' | 'failed';
-        archivePath?: string;
-        summaryPath?: string;
-        error?: string;
-      };
-
-      if (liveEvent.status === 'running') {
-        upsertLiveStatusMessage(
-          'live-run-status',
-          liveEvent.archivePath ? `Compacting earlier history… ${liveEvent.archivePath}` : 'Compacting earlier history…',
-          { pending: true, streaming: false },
-        );
-        return;
-      }
-
-      if (liveEvent.status === 'failed') {
-        upsertLiveStatusMessage(
-          'live-run-status',
-          liveEvent.error ? `Compaction failed: ${liveEvent.error}` : 'Compaction failed.',
-          { pending: false, streaming: false },
-        );
-        return;
-      }
-
-      if (liveEvent.status === 'finished') {
-        upsertLiveStatusMessage(
-          'live-run-status',
-          liveEvent.summaryPath ? `Compaction finished. Summary: ${liveEvent.summaryPath}` : 'Compaction finished.',
-          { pending: false, streaming: false },
-        );
-        return;
-      }
-
-      if (liveEvent.type === 'loop.started') {
-        setRunInFlight(true);
-        upsertLiveStatusMessage('live-run-status', 'Run started…', { pending: true, streaming: true });
-        return;
-      }
-
-      if (liveEvent.type === 'tool.calling' && typeof liveEvent.tool === 'string') {
-        upsertLiveStatusMessage(
-          'live-run-status',
-          `Working… running ${liveEvent.tool}${typeof liveEvent.step === 'number' ? ` (step ${liveEvent.step})` : ''}`,
-          { pending: true, streaming: true },
-        );
-        return;
-      }
-
-      if (liveEvent.type === 'tool.completed' && typeof liveEvent.tool === 'string') {
-        upsertLiveStatusMessage(
-          'live-run-status',
-          `${liveEvent.tool} finished${typeof liveEvent.durationMs === 'number' ? ` in ${Math.round(liveEvent.durationMs)}ms` : ''}`,
-          { pending: false, streaming: false },
-        );
-        return;
-      }
-
-      if (liveEvent.type === 'trace' && liveEvent.event) {
-        const traceEvent = liveEvent.event;
-        if (traceEvent.type === 'memory.maintenance_started') {
-          setMemoryUpdating(true);
-          upsertLiveStatusMessage(
-            'live-run-status',
-            `Memory updating… ${traceEvent.candidateIds.length} candidate${traceEvent.candidateIds.length === 1 ? '' : 's'}`,
-            { pending: true, streaming: false },
-          );
-          return;
-        }
-
-        if (traceEvent.type === 'memory.maintenance_finished') {
-          setMemoryUpdating(false);
-          upsertLiveStatusMessage(
-            'live-run-status',
-            traceEvent.summary ? `Memory updated. ${traceEvent.summary}` : 'Memory updated.',
-            { pending: false, streaming: false },
-          );
-          void refresh({ silent: true });
-          return;
-        }
-
-        if (traceEvent.type === 'memory.maintenance_failed') {
-          setMemoryUpdating(false);
-          upsertLiveStatusMessage(
-            'live-run-status',
-            traceEvent.error ? `Memory update failed: ${traceEvent.error}` : 'Memory update failed.',
-            { pending: false, streaming: false },
-          );
-          return;
-        }
-
-        if (traceEvent.type === 'tool.approval_requested') {
-          void fetchPendingSessionApproval(sessionId).then((approval) => setPendingApproval(approval));
-          upsertLiveStatusMessage(
-            'live-run-status',
-            `Approval requested for ${traceEvent.call.tool}${typeof traceEvent.step === 'number' ? ` (step ${traceEvent.step})` : ''}`,
-            { pending: true, streaming: false },
-          );
-          return;
-        }
-
-        if (traceEvent.type === 'tool.approval_resolved') {
-          setPendingApproval(null);
-          upsertLiveStatusMessage(
-            'live-run-status',
-            `Approval ${traceEvent.approved ? 'granted' : 'denied'} for ${traceEvent.call.tool}${traceEvent.reason ? ` — ${traceEvent.reason}` : ''}`,
-            { pending: false, streaming: false },
-          );
-          return;
-        }
-
-        if (traceEvent.type === 'tool.fallback') {
-          upsertLiveStatusMessage(
-            'live-run-status',
-            `Fallback: ${traceEvent.fromCall.tool} → ${traceEvent.toCall.tool}`,
-            { pending: true, streaming: false },
-          );
-          return;
-        }
-      }
-
-      if (liveEvent.type === 'assistant.stream' && typeof liveEvent.text === 'string') {
-        upsertLiveAssistantMessage(liveEvent.text, Boolean(liveEvent.done));
-        if (liveEvent.done) {
-          removeLiveStatusMessage('live-run-status');
-        }
-        return;
-      }
-
-      if (liveEvent.type === 'loop.finished') {
-        setRunInFlight(false);
-        removeLiveStatusMessage('live-run-status');
-        void refresh();
-      }
-    });
-
-    return () => {
-      cancelled = true;
-      if (sessionUpdateTimeout !== undefined) {
-        window.clearTimeout(sessionUpdateTimeout);
-      }
-      unsubscribe();
-    };
-  }, [
-    removeLiveStatusMessage,
+  useSessionDetailSubscription({
     selectedSessionId,
-    upsertLiveAssistantMessage,
-    upsertLiveStatusMessage,
-  ]);
-
-  useEffect(() => {
-    const latestTurnId = sessionDetail?.turns.at(-1)?.id;
-    if (!sessionDetail) {
-      setSelectedTurnId(undefined);
-      return;
-    }
-    if (!latestTurnId) {
-      setSelectedTurnId(undefined);
-      return;
-    }
-    if (!selectedTurnId || !sessionDetail.turns.some((turn) => turn.id === selectedTurnId)) {
-      setSelectedTurnId(latestTurnId);
-    }
-  }, [selectedTurnId, sessionDetail]);
+    setSessionDetail,
+    setSessionDetailLoading,
+    setSessionDetailError,
+    setRunInFlight,
+    setMemoryUpdating,
+    setPendingApproval,
+    onSessionsChanged,
+    liveMessages,
+  });
 
   const activeSession = useMemo(
     () => sessions?.find((session) => session.id === selectedSessionId),
     [selectedSessionId, sessions],
   );
-  const selectedTurn = useMemo(
-    () => sessionDetail?.turns.find((turn) => turn.id === selectedTurnId) ?? sessionDetail?.turns.at(-1),
-    [selectedTurnId, sessionDetail],
-  );
 
-  useEffect(() => {
-    if (!sessionDetail?.id || !selectedTurnId) {
-      setTurnReview(null);
-      setTurnReviewError(undefined);
-      return;
-    }
+  const {
+    selectedTurn,
+    turnReview: loadedTurnReview,
+    turnReviewLoading: loadedTurnReviewLoading,
+    turnReviewError: loadedTurnReviewError,
+  } = useSessionTurnReview({
+    sessionDetail,
+    selectedTurnId,
+    setSelectedTurnId,
+    turnReview,
+    setTurnReview,
+    turnReviewLoading,
+    setTurnReviewLoading,
+    turnReviewError,
+    setTurnReviewError,
+  });
 
-    const sessionId = sessionDetail.id;
-    const turnId = selectedTurnId;
-    let cancelled = false;
-    setTurnReviewLoading(true);
-
-    async function refresh() {
-      try {
-        const next = await fetchChatTurnReview(sessionId, turnId);
-        if (!cancelled) {
-          setTurnReview(next);
-          setTurnReviewError(undefined);
-        }
-      } catch (refreshError) {
-        if (!cancelled) {
-          setTurnReviewError(refreshError instanceof Error ? refreshError.message : String(refreshError));
-        }
-      } finally {
-        if (!cancelled) {
-          setTurnReviewLoading(false);
-        }
-      }
-    }
-
-    void refresh();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [selectedTurnId, sessionDetail?.id, selectedTurn?.traceFile]);
-
-  const resolveApproval = useCallback(async (approved: boolean) => {
-    if (!selectedSessionId || !pendingApproval) {
-      return;
-    }
-
-    try {
-      await resolvePendingSessionApproval(
-        selectedSessionId,
-        approved,
-        approved ? 'Approved in web control plane' : 'Denied in web control plane',
-      );
-      setPendingApproval(null);
-      notify?.({
-        title: approved ? 'Approval granted' : 'Approval denied',
-        body: pendingApproval.tool,
-        tone: approved ? 'success' : 'info',
-      });
-    } catch (error) {
-      notify?.({
-        title: 'Approval failed',
-        body: error instanceof Error ? error.message : String(error),
-        tone: 'error',
-      });
-    }
-  }, [notify, pendingApproval, selectedSessionId]);
-
-  const createSession = useCallback(async () => {
-    setCreatingSession(true);
-    setSessionNotice('Creating a new session…');
-    try {
-      const created = await createChatSession();
-      setSelectedSessionId(created.id);
-      setSessionDetail(created);
-      setSelectedTurnId(undefined);
-      setTurnReview(null);
-      setPendingApproval(null);
-      setRunInFlight(false);
-      setSendPromptError(undefined);
-      setSessionDetailError(undefined);
-      setSessionNotice(`Created ${created.name}. Ready for a fresh prompt.`);
-      notify?.({
-        title: 'Session created',
-        body: `${created.name} is ready for a fresh prompt.`,
-        tone: 'success',
-      });
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      setSessionNotice(message);
-      notify?.({
-        title: 'Session creation failed',
-        body: message,
-        tone: 'error',
-      });
-    } finally {
-      setCreatingSession(false);
-    }
-  }, [notify, setSelectedSessionId]);
-
-  const sendPrompt = useCallback(async (prompt: string) => {
-    const trimmed = prompt.trim();
-    if (!selectedSessionId || !trimmed || sendingPrompt) {
-      return;
-    }
-
-    setSendingPrompt(true);
-    setRunInFlight(true);
-    setSendPromptError(undefined);
-    appendPendingUserTurn(trimmed);
-    try {
-      const result = await sendChatSessionPrompt(selectedSessionId, trimmed);
-      setSessionDetail(result.session);
-      const latestTurnId = result.session?.turns.at(-1)?.id;
-      if (latestTurnId) {
-        setSelectedTurnId(latestTurnId);
-      }
-      setRunInFlight(false);
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      setSendPromptError(message);
-      setRunInFlight(false);
-      notify?.({
-        title: 'Send failed',
-        body: message,
-        tone: 'error',
-      });
-    } finally {
-      setSendingPrompt(false);
-    }
-  }, [appendPendingUserTurn, notify, selectedSessionId, sendingPrompt]);
-
-  const continueSession = useCallback(async () => {
-    if (!selectedSessionId || sendingPrompt || runInFlight) {
-      return;
-    }
-
-    setSendingPrompt(true);
-    setRunInFlight(true);
-    setSendPromptError(undefined);
-    upsertLiveStatusMessage('live-run-status', 'Continuing from the current transcript…', { pending: true, streaming: true });
-
-    try {
-      const result = await continueChatSession(selectedSessionId);
-      setSessionDetail(result.session);
-      const latestTurnId = result.session?.turns.at(-1)?.id;
-      if (latestTurnId) {
-        setSelectedTurnId(latestTurnId);
-      }
-      setPendingApproval(await fetchPendingSessionApproval(selectedSessionId));
-      setRunInFlight(false);
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      setSendPromptError(message);
-      setRunInFlight(false);
-      notify?.({
-        title: 'Continue failed',
-        body: message,
-        tone: 'error',
-      });
-      removeLiveStatusMessage('live-run-status');
-      setSessionDetail((current) => {
-        if (!current) {
-          return current;
-        }
-        return {
-          ...current,
-          messages: current.messages.filter((message) => message.id !== 'live-assistant'),
-        };
-      });
-    } finally {
-      setSendingPrompt(false);
-    }
-  }, [notify, removeLiveStatusMessage, runInFlight, selectedSessionId, sendingPrompt, upsertLiveStatusMessage]);
-
-  const cancelSessionRun = useCallback(async () => {
-    if (!selectedSessionId || !runInFlight) {
-      return;
-    }
-
-    try {
-      const result = await cancelChatSession(selectedSessionId);
-      setRunInFlight(false);
-      setPendingApproval(null);
-      removeLiveStatusMessage('live-run-status');
-      setSendPromptError(undefined);
-      setSessionDetail((current) => {
-        if (!current) {
-          return current;
-        }
-        return {
-          ...current,
-          messages: current.messages.filter((message) => message.id !== 'live-user' && message.id !== 'live-assistant'),
-        };
-      });
-      notify?.({
-        title: result.cancelled ? 'Run cancelled' : 'No active run to cancel',
-        body: result.cancelled ? 'The active session run was interrupted.' : undefined,
-        tone: result.cancelled ? 'success' : 'info',
-      });
-    } catch (error) {
-      notify?.({
-        title: 'Cancel failed',
-        body: error instanceof Error ? error.message : String(error),
-        tone: 'error',
-      });
-    }
-  }, [notify, removeLiveStatusMessage, runInFlight, selectedSessionId]);
-
-  const updateSessionSettings = useCallback(async (settings: { model?: string; driftEnabled?: boolean }) => {
-    if (!selectedSessionId || runInFlight || sendingPrompt) {
-      return;
-    }
-
-    try {
-      const updated = await updateChatSessionSettings(selectedSessionId, settings);
-      setSessionDetail(updated);
-      notify?.({
-        title: 'Session settings updated',
-        body: [
-          settings.model ? `model ${settings.model}` : undefined,
-          typeof settings.driftEnabled === 'boolean' ? `drift ${settings.driftEnabled ? 'on' : 'off'}` : undefined,
-        ].filter(Boolean).join(', '),
-        tone: 'success',
-      });
-    } catch (error) {
-      notify?.({
-        title: 'Settings update failed',
-        body: error instanceof Error ? error.message : String(error),
-        tone: 'error',
-      });
-    }
-  }, [notify, runInFlight, selectedSessionId, sendingPrompt]);
+  const {
+    resolveApproval,
+    createSession,
+    sendPrompt,
+    continueSession,
+    cancelSessionRun,
+    updateSessionSettings,
+  } = useSessionMutations({
+    selectedSessionId,
+    pendingApproval,
+    sendingPrompt,
+    runInFlight,
+    notify,
+    setSelectedSessionId: setRawSelectedSessionId,
+    setSelectedTurnId,
+    setSessionDetail,
+    setSessionDetailError,
+    setSendingPrompt,
+    setSendPromptError,
+    setTurnReview,
+    setPendingApproval,
+    setRunInFlight,
+    setCreatingSession,
+    setSessionNotice,
+    liveMessages,
+  });
 
   return {
     activeSession,
@@ -709,28 +157,8 @@ export function useSessionsScreenState(
     selectedTurnId,
     setSelectedTurnId,
     selectedTurn,
-    turnReview,
-    turnReviewLoading,
-    turnReviewError,
-  };
-}
-
-function mergeTransientMessages(current: ChatSessionDetail | null, next: ChatSessionDetail | null): ChatSessionDetail | null {
-  if (!current || !next || current.id !== next.id) {
-    return next;
-  }
-
-  const transientMessages = current.messages.filter((message) => message.id.startsWith('live-'));
-  if (!transientMessages.length) {
-    return next;
-  }
-
-  const nextMessageIds = new Set(next.messages.map((message) => message.id));
-  return {
-    ...next,
-    messages: [
-      ...next.messages,
-      ...transientMessages.filter((message) => !nextMessageIds.has(message.id)),
-    ],
+    turnReview: loadedTurnReview,
+    turnReviewLoading: loadedTurnReviewLoading,
+    turnReviewError: loadedTurnReviewError,
   };
 }
